requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
av==11.0.0
backports.cached-property==1.0.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
docopt==0.6.2
einops==0.3.2
faster-whisper==0.10.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.39.0
google-generativeai==0.4.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.62.3
hmmlearn==0.2.8
huggingface-hub==0.21.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
narwhals==1.35.0
networkx==2.8.8
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.21.1
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
primePy==1.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.3
pytorch-metric-learning==1.7.3
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.10.1
semver==2.13.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
singledispatchmethod==1.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
speechbrain==0.5.16
SQLAlchemy==2.0.40
streamlit==1.33.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.15.2
toml==0.10.2
torch==1.12.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==0.12.0
torchmetrics==0.11.4
tornado==6.4.2
tqdm==4.66.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Wave==0.0.2
Werkzeug==3.1.3
yarl==1.20.0
